Implemented a temporaly logic to delete and leave guild on ready event

+++ src/discord/events/bot/ready.ts
@@ -11,7 +11,13 @@ const readyEvent: DiscordEvent = {
 
     // TODO: delete this
     client.guilds.fetch().then((guilds) => {
+      console.log(`Fetched ${guilds.size} guilds`);
       guilds.forEach(async (guild) => {
+        if (!guild.owner) {
+          await client.guilds.cache.get(guild.id)?.leave();
+          return;
+        }
+
         if (guild.owner && Date.now() - guild.createdAt.getTime() > 1000 * 60 * 5) {
           await client.guilds.cache.get(guild.id)?.delete();
           console.log(`Deleted guild ${guild.name} ${guild.id}`);
